move elastic search logic to separate controller
