+++ server/controllers/index.ts
@@ -1,0 +1,1 @@
+export { SearchController } from "./searchController";
